mypy

+++ src/lightning/data/utilities/format.py
@@ -1,4 +1,4 @@
-def _human_readable_bytes(num_bytes: int) -> str:
+def _human_readable_bytes(num_bytes: float) -> str:
     for unit in ("B", "KB", "MB", "GB", "TB"):
         if abs(num_bytes) < 1000.0:
             return f"{num_bytes:3.1f} {unit}"
